Prevent the scanner from recursively peeking on error.
Plex uses the position() method when constructing an error object, but
if an error occurs while we're already peeking then we end up trying to
peek again.

This is a bit of a gross hack but it works to preserve the expected
behavior of position() in this scenario.

As a bonus, we also now catch the UnrecognizedInput error from Plex and
promote it to a CompilerError subclass so our logging system can
deal with it nicely.