Special AST nodes for resolved values and symbol storages.
These are never emitted by the parser, but the interpreter will emit these
once it's bound source expressions to real symbols and values, and these
will be used to emit the right references during code generation.

Therefore these AST nodes are only valid in the code generation tree, not
in the parse tree.