Allow log lines to carry additional detail lines.
This facility is intended to carry additional details that may help the
user fix a problem but are not necessary to understand what sort of
problem has occurred. Compile UIs may or may not show these additional
details, or the details might be initially hidden until the user
explicitly requests them.