Make ValueExpr and SymbolExpr take positions rather than source_nodes.
Initially I thought it would be useful to be able to refer back to the
source for these, but we've never done that so far except in tests.
Instead this is a vote for consistency with all of the other node types,
and with the move towards using positions as the common currency for
retaining source references to include in error messages.