Correct and simplify the handling of while loops.
Previously the rules for while loops were complicated, and to make matters
worse we made some incorrect assumptions in our handling of generating
a while loop for runtime.

This change simplifies the rules for while loops so that they are only
evaluated at compile time if they can be completely removed from the
runtime code. If any part of the while loop cannot be handled at compile
time, we use the "force runtime" interpreter feature to produce a
runtime version of the loop.