Require symbols to be initialized.
It is mandatory for all symbols to be initialized before their containing
scope exits. At that point we'll bake in the symbol's type to be
henceforth unchanging, so that it isn't subject to being "merged" into
any parent data state, where it would get corrupted since the parent
state doesn't know the symbol's type.