Support retrieving a flattened set of symbols from a symbol table.
This introduces the concept of a "disconnected" child symbol table, which
is one that should not participate in flattening. The use-case for this
is cases where a child symbol table represents a separate execution
context like a function, where we expect the function to be able to reach
into the global scope but not for the locals to roll up into the global
scope.