Further ad-hoc progress on the preprocessor.
The basics of the preprocessor are now working, but a bit missing part
is the merge operation on the DataState class which is supposed to create
an entry data state based on the exit states of the predecessor blocks.
Currently this is not done and so the preprocessor's type analysis is
incorrect and incomplete.