Keep track of type equivalences that arise in unification.
This allows us to deal with the situation where what was previously two
separate types unifies into one single type. After unification we keep
track of the equivalences in a table, and then on retrieval we
opportunistically flatten the equivalence map so that the caller will see
the correctly-unified type.

Along the way the internal state for TypeTable became a pair of weakref
dictionaries, so we don't hold on to state from previously-visited
functions for the entire compilation process.