Correct implementation of TypeTable equality.
Since type variables are different each time we run, we need to be careful
to compare these as equal regardless of their identity. When we talk about
changes to the type inference we really mean changes to what we know, not
changes to what we *don't* know.

This allows type inference to converge, where previously it was
impossible.