Thread-local context object.
Rather than passing around all of the state info that is effectively
dynamically-scoped during preprocessing, instead we'll keep it in a
thread-local storage and provide a context manager to manage it.

This is a little unusual but it has the advantage that our operand objects
will get to behave like normal objects even though part of their data
comes from locations that vary during processing.