First run at LLVM code generation.
The architecture's still a little odd here but this represents the first
time we've been able to go all the way from Alamatic source to LLVM
assembly, at least for a very contrived program with basic looping and
branching using integers.

Not too happy with how this spread LLVM concepts all over the intermediate
layer, but it'll do for now. Perhaps there will be some refactoring later.