Experimental use of the LLVM optimizer in alac.
This is obviously not how this will finally be implemented, but more of
an experiment to see how the pass manager interface works, and how we
can generate native code. For the moment, we can only generate native
code for the host platform, so more investigation is required to
generate code for ARM cross-compilation.