FunctionTemplate type implementation skeleton.
Once the preprocessor is more fleshed out, we should eradicate all
mentions of these before we do final codegen, but when we're doing
compile-time function evaluation we permit these to be passed around for
metaprogramming purposes. Thus this represents our first example of
exposing a PyObject* pointer into LLVM land so that we can fish it back
out later if e.g. a compile-time function returns one of these.