Tear everything back to just the scanner and beginnings of parser.
I've reached the conclusion that there are some fundamental design errors
in the design of the intermediate language that were causing the analysis
and code generation steps to be far more complicated than necessary.

Removing everything after the parser also gives an opportunity to give
the parser a bit of a spring clean.

This also eliminates the ill-conceived "CompilerState" object in favor of
in-band error reporting via some special error-marking AST nodes. This
allows parsing errors to be handled only when the relevant AST subtree is
processed, meaning we won't report errors in parts of libraries that
aren't used by the current program.