Introduce a new diagnostics system.
Rather than having random exception types sprinkled all over the codebase,
we create a central registry of diagnostics. This allows them to all
share the same mechanism and should make it easier to do things like
allowing the user to suppress certain kinds of diagnostic, or to produce
localized versions of the diagnostic messages once they've stabilized a
little bit.