Refactor the parser to be a class.
Previously we had each parser function accepting a state and a scanner
from the caller, and all of the calls shared the same values for these,
so it seems more natural to wrap the whole thing up in a class that can
maintain that state.

The 'state' parameter is actually no longer needed -- the CompileState
object was removed in an earlier commit anyway -- so the Parser just
wraps a Scanner.