Rework how literals are handled.
Rather than having multiple different operand types on the operations, we
just force operations to take registers and provide an operation for
loading a literal value into a register.

This simplifies the model a bunch since now the handling of literals is
completely separated from the handling of operations. When we later lower
to LLVM IR we will collapse these back into direct constant operands via
constant folding.

This also includes the start of a new value/type system, and a stub
implementation of the constant number type since I needed something to
instantiate to flesh out the new model for literals.