Change octal number syntax to 0o rather than just 0 prefix.
C's 0 prefix for octal is widely considered to be a mistake that trips
people up. An explicit "0o" prefix goes with "0x" and "0b" and makes it
much less likely that octal will accidentally be introduced.

This also fixes a pre-existing bug in the parser where fractions had to
be written as e.g. ".0" because "0.0" would cause us to try to parse
as octal.