Stub out the beginnings of the LLVM lowering code.
This module is responsible for turning Alamatic IR into LLVM IR, which
involves doing type inference, constant folding and the elimination of
basic blocks that depend on constant conditions that are false.

For now, it just generates an empty shell of an LLVM module with the local
variables defined, albeit with the wrong types.