Generate synthetic newline and EOF tokens at end of stream.
The higher-level layers will need to know the location of the end of the
file in order to emit usable diagnostics for things like a premature EOF,
so we emit it as a token to give us somewhere to put that information.

Also emit an extra newline at the end so that all statements are guaranteed
to end with a newline, and so that the EOF is always on a line of its own
and won't get confused with any real tokens.