Indent bytes occupy the same area as their spaces.
Previously the "bytes" associated with an indent token were set to
represent the number of *additional* columns of indentation that were
being added in the new block, but that causes the token's source range
to be misleading as its start location is anchored at the beginning
of the line.

Instead we'll just return the total number of spaces we're indenting,
causing the token to span the whole set of spaces from the start of line
to the first visible character.