ir: refactor and reorganize
The previous IR representation was rather cumbersome. This is the beginning
of a new one, modelled after the SSA IR used for Go.

This breaks the interface that the ast package uses to lower to IR, so
its tests fail as a result of this change. This will be addressed in a
later commit.